// src/integrations/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
// ...
      messages: {
        Row: {
          color: string | null
          created_at: string
          gift_data: Json | null
          id: string
          is_whisper: boolean | null
          message: string
          speech: string | null
          user_id: string | null
          username: string
          whisper_target_id: string | null
        }
        Insert: {
          color?: string | null
          created_at?: string
          gift_data?: Json | null
          id?: string
          is_whisper?: boolean | null
          message: string
          speech?: string | null
          user_id?: string | null
          username: string
          whisper_target_id?: string | null
        }
        Update: {
          color?: string | null
          created_at?: string
          gift_data?: Json | null
          id?: string
          is_whisper?: boolean | null
          message?: string
          speech?: string | null
          user_id?: string | null
          username?: string
          whisper_target_id?: string | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}